+++ acme/Storage.py
@@ -573,7 +573,8 @@ class TinyDBBinding(object):
 										'exc' : subscription.exc,
 										'ln'  : subscription.ln,
 										'nus' : subscription.nu,
-										'bn'  : subscription.bn
+										'bn'  : subscription.bn,
+										'exc' : subscription.exc,
 									}, 
 									Query().ri == ri)	# type: ignore
 			return result is not None